Select root-console without assert and clean
When running HA tests in the serial terminal, the root console can be
filled by console messages which remove the root prompt. The
`ha/fencing` module needs to run in the root console, and the first
thing it does is to select it, but if there is no root prompt there, it
will fail on a needle.

This commit introduces a workaround such that the console is first
selected with `await_console => 0` to skip the needle check, then the
console is cleared and a `ret` key is sent to cause it to show the
prompt, before asserting the `root-console` needle.